Misc changes
- Moved LabelUnit struct to LabelContainer header
- Label does not own any textbuffer now!! when required its passed by reference!
        - ownership with fonstash and maptile.
